Fix the NewTicker race and add regression tests.
This fixes the race described by #17. The test provided in that issue is
added along with another similar test that also reproduced the racy
behavior.

The fix is to synchronously create the wait channel for the ticker
rather that lazily create it. This must also occur on each tick: the
next waiter must be initialized before the tick is sent.

Fixes https://github.com/jonboulle/clockwork/issues/17