Fix a slight race in a ticker test.
This is an incredibly unlikely race, but it shows up when running the
test 30k+ times on my system. The problem is that it's possible for:
* the fake clock to unblock the ticker
* the test attempt to read from the channel before the ticker sends

This is more likely in the new implementation than the old
implementation before there is slightly more logic between the next
tick reception and the tick send.