Merge the concepts of sleeper and timer
A timer with an event in the future is essentially the same thing as a
sleeper.  So this commit merges the two, using the `fakeTimer` structure
where `sleeper` had been used before.  Since management of sleepers is
performed while holding the lock of the fake clock, this means that the
basic timer operations are now synchronized using these locks as well, which
should go a long way to making all operations easier to reason about without
messing around with atomics.

Fixes https://github.com/jonboulle/clockwork/issues/42.