Switch ticker implementation to timers as a basis
This avoids the separate goroutine and several levels of indirection.

It replaces the previous approach of skipping a number of events with an
approach that at least conceptually delivers each tick to the channel if
possible without blocking.  That is easier to reason about and will more
clearly ensure correct ordering of events, as demonstrated by the new test
case which had a high probability to fail with the previous implementation.