Convert FakeClock to a struct.
This change ensures future modifications to FakeClock will not break users.

However, this change does break current users who use FakeClock as an argument,
return value, or embed it in a struct or interface.

With limtied exception, build errors can be fixed by running:
```
$ grep -rl 'clockwork.FakeClock' /path/to/code_base | \
    xargs sed -i 's/clockwork.FakeClock/*clockwork.FakeClock/g'
```

Other errors and their fixes are:
- When returning a pointer to concrete type as a clockwork.FakeClock: return
  the pointer to concrete type.
- When using FakeClock in an interface composition: relace with the functions
  required by the interface.
- When using FakeClock as an argument, replace with concerete type or define a
  new interface.

The anternative approach would have been to use type alaises. However, this
does not provide a benefit over direct replacement, because the migration is
more complicated, harder to execute, needs to be done to the same code points,
and results in runtime failures (whereas this approach can be complie-time
checked).